Add selectable table js
This commit adds the ability to select more rows on admin resource
tables.
To use this feature, add the `selectable-table` CSS class to the
selected table and the CSS class `selectable` to the checkbox, which
identifies the record id.

It is possible to add a checkbox named `select-all` to check/uncheck all
the elements.

Adding the selector CSS class `selectable-label` on a node HTML it will
be rendered a text depending on the selected rows number.

An example of usage is provided at `/admin/style_guide` under
`Building Tables` section.