Add selectable table js
This commit adds the ability to select more rows on admin resource
tables.
To use this feature, add the `selectable-table` CSS class to the
selected table and the CSS class `selectable` to the checkbox, which
identifies the record id.

It is possible to add a checkbox named `select-all` to check/uncheck all
the elements.

An example of its usage is provided at `/admin/style_guide` under
`Building Tables` section.